% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songsim.R
\name{songsim}
\alias{songsim}
\title{Calculating and Visualizing a "Song Similarity" matrix}
\usage{
songsim(path = NULL, colorfulMode = FALSE, singleColor = "black",
  interactiveMode = FALSE, mainTitle = "", plotOptions = NULL,
  heatmaplyOptions = NULL, plotMatrix = TRUE)
}
\arguments{
\item{path}{Path to a txt file holding the song's lyrics}

\item{colorfulMode}{A boolean indicating whether to plot the matrix in color
(see @details), defaults to FALSE}

\item{singleColor}{What is the color to use for a full matrix cell (see @details)}

\item{interactiveMode}{A boolean indicating whether to plot the matrix in
interactive mode using the `heatmaply` package (see @details), defaults to FALSE}

\item{mainTitle}{plot's main title, defaults to an empty string}

\item{plotOptions}{Additional parameters for the base `plot` function
(non-interactive mode)}

\item{heatmaplyOptions}{Additional parameters for the `heatmaply` function
(non-interactive mode)}

\item{plotMatrix}{A boolean indicating whether to plot the matrix, defaults to TRUE}
}
\value{
a list containing:
\item{songMat}{a song similarity square matrix of dimensions no. of words x
no. of words, containing all 0 and 1}
\item{repetitiveness}{an experimental measure of repetitiveness which is the
mean of the upper triangular part of song matrix}
}
\description{
This function will calculate the "song similarity" of a given song's lyrics
stored in a text file. A cell (i, j) is filled if word i in a song's lyrics
is the same as word j. This method was first suggested by Colin Morris
(see URLs)
}
\details{
A cell (i, j) in a songsim square matrix is filled if word i in a song's lyrics
is the same as word j. This simple rule would create fascintaing visualizations
of songs lyrics as shown in the links below.

If \code{colorfulMode} is set to TRUE, each word which appears more than once will be
colored with its own color (currently only available with \code{interactiveMode} set to FALSE).

If \code{interactiveMode} is set to TRUE and the user has the \code{heatmaply} package
installed - the songsim matrix will be plotted using an interactive heatmap.

The \code{singleColor} corresponds to the color a full cell is colored with.
If in \code{colorfulMode} it would be the color of words which appear only once.
}
\note{
This function was only tested on typical pop songs or poems, it was not tested
on large pieces of texts
}
\examples{
path <- system.file("extdata", "formation.txt", package = "songsim")
songsim(path)
}
\references{
A blog post describing the package with more examples:
\url{http://giorasimchoni.com/2017/08/08/2017-08-08-lambada-the-mocap-package/}

A blog post by Colin Morris describing interesting "songsims":
\url{https://colinmorris.github.io/blog/weird-pop-songs}

An interactive React (JS) demo by Colin Morris with more examples:
\url{https://colinmorris.github.io/SongSim/}
}
